renamed experimental views, added volume control
