+++ crmetronome.ui/package.json
@@ -9,7 +9,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.0.2",
-    "react-scripts": "4.0.3",
+    "react-scripts": "^5.0.0",
     "react-strap": "0.0.1",
     "reactstrap": "^9.0.1",
     "sass": "^1.49.9",
