+++ crmetronome.ui/src/App/index.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import './App.css';
 import { BrowserRouter as Router } from 'react-router-dom';
 import '../styles/index.scss';
