+++ crmetronome.ui/src/data/Routes.js
@@ -1,20 +1,25 @@
+import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import Sound from '../views/Sound';
+import Home from '../views/Home.js';
+import SoundContext1 from '../views/SoundContext1';
 import SoundContext2 from '../views/SoundContext2';
-import SoundContext1 from '../views/SoundContext';
+import SoundContext3 from '../views/SoundContext3';
 
 
 const RouterComponent = () => (
   <div>
     <Routes>
-      <Route path='/sounds'
-      element={ <Sound /> }
+      <Route path='/'
+        element={ <Home /> }
       />
-      <Route path='/soundContext'
-      element={ <SoundContext1 /> }
+      <Route path='/soundContext1'
+        element={ <SoundContext1 /> }
       />
       <Route path='/soundContext2'
-      element={ <SoundContext2 /> }
+        element={ <SoundContext2 /> }
+      />
+      <Route path='/soundContext3'
+        element={ <SoundContext3 /> }
       />
     </Routes>
   </div>
