+++ crmetronome.ui/src/views/Home.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+
+const Home = () => {
+    return (
+            <div>Home Page</div>
+    );
+};
+
+export default Home;
