+++ crmetronome.ui/src/views/SoundContext1.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import BassDrum2 from '../Assets/WavSamples/Bass-Drum-2.wav'
 
-const Sound = () => {
+const SoundContext1 = () => {
   // const [song, setSong] = useState({});
 
 /*  useEffect(() => {
@@ -20,13 +20,13 @@ const Sound = () => {
     // create Oscillator node
   const oscillator = audioCtx.createOscillator();
   const gainNode = audioCtx.createGain();
+  gainNode.gain.value = 0.1;
+  gainNode.connect(audioCtx.destination);
   oscillator.type = 'square';
   oscillator.frequency.setValueAtTime(440, audioCtx.currentTime); // value in hertz
   oscillator.connect(gainNode);
-  oscillator.connect(audioCtx.destination);
   audioCtx.suspend();
   oscillator.start();
-  gainNode.gain.value = 0.1;
     const handleStartStop = () => {
       if (audioCtx.state === 'suspended') {
         audioCtx.resume();
@@ -37,7 +37,7 @@ const Sound = () => {
   
   return (
   <>
-  <div>Sounds</div>
+  <div>SoundContext1</div>
   <div>
     <audio controls>
       <source src={BassDrum2} />
@@ -49,4 +49,4 @@ const Sound = () => {
   );
 };
 
-export default Sound;
+export default SoundContext1;
